Add sequences, chords and simple visualization
This adds four parts that play a sequence of tones and two that play
chords. Sequences and chords are configurable.

The configuration is long and some of that could be moved to controls,
but a lot of setting up Tone.js stuff is abstracted away into a handful
of modules now.

The Chord module is particularly neat. It takes advantage of music-fns
so that we can instantiate a Chord instance with a root note and call
getter methods on it to get major, minor and minorAugmented chord notes
back, which we can then pass on to Tone.js to make a PolySynth playing
the chord. Adding more chord variations should be trivial.